+++ src/scalar/utf16.h
@@ -20,7 +20,7 @@ inline simdutf_warn_unused bool validate(const char16_t *buf, size_t len) noexce
         if(pos + 1 >= len) { return false; }
         uint16_t diff = uint16_t(word - 0xD800);
         if(diff > 0x3FF) { return false; }
-        uint16_t next_word = big_endian ? (data[pos + 1] >> 8 | data[pos + 1] << 8) : data[pos + 1];
+        uint16_t next_word = big_endian ? uint16_t((data[pos + 1] >> 8) | (data[pos + 1] << 8)) : data[pos + 1];
         uint16_t diff2 = uint16_t(next_word - 0xDC00);
         if(diff2 > 0x3FF) { return false; }
         pos += 2;
